+++ src/main/kotlin/com/github/gib/actions/GHomeAction.kt
@@ -1,23 +1,20 @@
 package com.github.gib.actions
 
+import com.github.gib.services.GivServiceSettings
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.DumbAware
 import com.intellij.ui.jcef.JBCefBrowser
 import javax.swing.Icon
 
-class GBackAction(private val jbCefBrowser: JBCefBrowser, icon : Icon) : AnAction(icon), DumbAware {
+class GHomeAction(private val jbCefBrowser: JBCefBrowser, icon: Icon) : AnAction(icon), DumbAware {
 
 
     override fun update(e: AnActionEvent) {
-        if (!jbCefBrowser.cefBrowser.canGoBack()) {
-            e.presentation.isEnabled = false
-            return
-        }
-        super.update(e)
+        e.presentation.isEnabled = true
     }
 
     override fun actionPerformed(e: AnActionEvent) {
-        jbCefBrowser.cefBrowser.goBack()
+        jbCefBrowser.loadURL(GivServiceSettings.instance().getLastSaveHomePage())
     }
 }