+++ src/main/kotlin/com/github/gib/actions/GRefreshAction.kt
@@ -2,10 +2,11 @@ package com.github.gib.actions
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAware
 import com.intellij.ui.jcef.JBCefBrowser
 import javax.swing.Icon
 
-class GRefreshAction(private val jbCefBrowser: JBCefBrowser, icon: Icon) : AnAction(icon) {
+class GRefreshAction(private val jbCefBrowser: JBCefBrowser, icon: Icon) : AnAction(icon), DumbAware {
 
 
     override fun update(e: AnActionEvent) {