+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-
 ## [Unreleased]
+### Added
+- Back ,Forward, Refresh and home Buttons
+- Default Home Page in Settings
+- Url Text field with history
+- Support IJ 211
 ## [1.0.7]
 ### Added
 - Back ,Forward and Refresh Buttons